Provisioning changes (2)
- Added missing DB creation statement
- ExtJS won't be loaded if we put it manually

+++ provision.sh
@@ -31,16 +31,18 @@ apt-get install -y unzip
 
 export DEBIAN_FRONTEND=noninteractive
 apt-get -q -y install mysql-server
+mysql -u root -e 'CREATE DATABASE WorkPlace;'
 mysqladmin -u root password 123456
-mysql -u root -p'123456' WorkPlace < dump.sql
+mysql -u root -p'123456' WorkPlace < /vagrant/dump.sql
 
-cd WorkPlace/src/main/webapp
-mkdir extjs
-cd extjs
-
-wget https://olex-secure.openlogic.com/content/private/5e6a6f0815e830bba705e79e4a0470fbee8a5880//olex-secure.openlogic.com/ext-5.1.0-gpl.zip
-unzip ext-5.1.0-gpl.zip
-rm ext-5.1.0-gpl.zip
+if [ ! -d /vagrant/WorkPlace/src/main/webapp/extjs ]; then
+    cd WorkPlace/src/main/webapp
+    mkdir extjs
+    cd extjs
+    wget https://olex-secure.openlogic.com/content/private/5e6a6f0815e830bba705e79e4a0470fbee8a5880//olex-secure.openlogic.com/ext-5.1.0-gpl.zip
+    unzip ext-5.1.0-gpl.zip
+    rm ext-5.1.0-gpl.zip
+fi
 
 cd /vagrant/WorkPlace
 mvn tomcat7:run
